Create Day_8
QUESTIONS 1.

Peter has been given an assignment to implement the code to convert the temprature from FORENHITE to CELCIUS and vice versa. Kindly help him to implement the code such that user can choose the conversion type

FORENHITE to CELCIUS CELCIUS to FORENHITE and then enter the value of temperature to be converted and display the temptrature in the appropriate format.

If user enters the choice any thing else then print INVALID CHOICE

Input Format

First line will take the choice N (integer) as input i.e. 1 or 2 Second line will take the temperature (double) T as input

Constraints

N >0 and N < 3

Output Format

Prints the double value i.e. the temperature

Sample Input 0

1
40.0
Sample Output 0

4.44
Sample Input 1

2
15.5
Sample Output 1

59.9


QUESTIONS 2.

A student will not be allowed to sit in exam if his/her attendence is less than 75%. Take following input from user Number of classes held Number of classes attended. And print percentage of class attended Is student is allowed to sit in exam or not.

Input Format

Number of classes held Number of classes attended.

Constraints

Number of classes held >= Number of classes attended. and both must be Positive integer

Output Format

percentage of class attended Is student is allowed to sit in exam or not.

Sample Input 0

100
80
Sample Output 0

80
Yes
Explanation 0

Explanation 0

Number of classes held : 100 Number of classes attended : 80 And Output percentage of class attended : 80 Is student is allowed to sit in exam or not : Yes

Sample Input 1

50
34
Sample Output 1

68
No
Explanation 1

Explanation 1

Number of classes held : 50 Number of classes attended : 34 And Output percentage of class attended : 68 Is student is allowed to sit in exam or not : No