Create Day_12
Questions 1.
Vijay and Gaurav are two friends. They are playing a game to check that whether the first word is similar to second or not. For example first word is "loop" and second is "lloooopp". These words will be considered similar as characters are same and in first word 'l' is occuring 1 time, 'o' is 2 times and 'p' is 1 time i.e occurance of 'l'= occurance of 'p' and occurance of 'o' is double than the occurance of 'l' and 'p'. Help them using program.

Input Format

First line will contain two space seperated words.

Constraints

Input will be two space seperated words only.

Output Format

Similar/ Not Similar

Sample Input 0

loop llooooopp
Sample Output 0

Not Similar
Sample Input 1

cool looc
Sample Output 1

Similar

Questions 2.

Alex is very fond of pattern making so he is writing a program to print the pattern of following type .He gives an input of int type which represents the total number of rows and the pattern will be printed up to the entered row size but if Alex enter the negative number ,show “Invalid Input”. Pattern for 2 rows: & && &&&

Input Format

Program should take the number of rows as input.

Constraints

Entered Number of rows should be positive int value

Output Format

If input is of positive number display the pattern but if input is a negative number or zero, display the message “Invalid Input”.

Sample Input 0

-5
Sample Output 0

Invalid Input
Sample Input 1

2
Sample Output 1

&
&&

Questions 2.
Alex is very fond of pattern making so he is writing a program to print the pattern of following type .He gives an input of int type which represents the total number of rows and the pattern will be printed up to the entered row size but if Alex enter the negative number ,show “Invalid Input”. Pattern for 2 rows: & && &&&

Input Format

Program should take the number of rows as input.

Constraints

Entered Number of rows should be positive int value

Output Format

If input is of positive number display the pattern but if input is a negative number or zero, display the message “Invalid Input”.

Sample Input 0

-5
Sample Output 0

Invalid Input
Sample Input 1

2
Sample Output 1

&
&&
Submissions: 190
Max Score: 5
Difficulty: Medium
Rate This Challenge:

    
More