Create Day_13
Questions 1.
Sameer wants to develop a program for ATM. Help him by constructing the program for the same. In the program if user withdraws amount upto 1000 then machine will dispence Rs. 100 notes only. Minimum number of notes should be dispenced by the machine. Notes of denomination of 100,200,500 and 2000 are available in machine.

Input Format

One integer value between 100-20000.

Constraints

Maximum withdrawl amount should be 20000, Minimum should be 100 and amount should be in the multiple of 100 only.

Output Format

Print the number of notes dispenced and their denomination.

Sample Input 0

1100
Sample Output 0

1 100 Notes
2 500 Notes
Sample Input 1

1550
Sample Output 1

Invalid Input

Questions 2.
Sajal and Ruhi are playing a game. Sajal is going to tell one number and one operation (Even, Odd, Prime and Factorial), Ruhi have to apply the operation and tell the output.

Input Format

Two space seperated numbers, first number on which need to apply operation and second number will be 1 for checking the number is even or not, 2 for checking number is odd or not, 3 for checking number is prime or not and 4 for calculating factorial of number.

Constraints

First number will be positive integer value. Second number can be in range of 1 to 4.

Output Format

Yes / No for first three operations and an integer value if operation four is performed.

Sample Input 0

5 1
Sample Output 0

NO
Sample Input 1

5 4
Sample Output 1

120