Create Day_19
Questions 1.
write a program that creates integer array of n elements, accepts the values of arrays and find sum of elements as an integer.

Input Format

Accepts n values from the user.

Constraints

All number should be integer values (Positive, negative and zero)

Output Format

diplays the sum of the array's elements as a single integer.

Sample Input 0

5
3 4 10 11 20
Sample Output 0

48

Questions 2.
Karan has been given a task by Head of Librarian to implement the enumeration of the Library working Day and timing. Which have the 3 attributes i.e. day and open timing and close timing of the library working hours. Details of the working hours are:

Day, open Timing, closed timing

Monday, 8.30, 5.30 Tuesday, 8.30, 5.30 . . Friday, 8.30, 5.30 Saturdays, 9.30, 4.30

Karan is supposed to write a main class in which the name of working days of library is taken as input as a day and timing of library should be displayed using switch statement. If the input given is not matching with the listed day, display Library is closed

Input Format

First line of the input reads the day

Constraints

NA

Output Format

if working day than Prints the open time and closed seperated by SPACE. else Prints Library is closed

Sample Input 0

Monday
Sample Output 0

8:30 5:30