Create Day_23
Questions 1.
Create a class Student with attributes name(String) and registrationNo(int) and a method show to display the attributes of the Student. Write a menu driven program with 2 choices to take the input of the student attributes and display the attributes.

Input Format

Your program should take the 2 choice as input. • If the first input will be 1 then also ask to user to enter the details of the student. If the entered registrationNo is negative, do not accept the input and prompt again to take new input until user will not entered the positive registrationNo. • If the first input will be 2 then display the default values of Student attributes separated by comma in the same line. • If the first input will be any other int value, display the message “Wrong Choice”.

Constraints

Student registrationNo should be positive value(excluding 0).

Output Format

Your program should display the attributes of Student or the message “Wrong Choice” as the User choice at the runtime.

Sample Input 0

1
John
10
Sample Output 0

John,10
Sample Input 1

2
Sample Output 1

null,0

Questions 2.
In some organization, January is the month of giving appraisal to employees, so HR department of that organization decided to give appraisal to only those employees whose employee id is more than 201. In that organization every employee is getting basic salary is 20,000INR and when employee id is greater than 201 then company decides to give appraisal of 1000INR to every employee. So as a programmer help that company to check who is eligible for getting incentive.

Input Format

In First Line, enter employee ID

In Second Line, enter employee name

In third Line, enter employee designation

Constraints

Check is on employee ID, and employee id must be of integer type.

Output Format

If employee id is greater than 201, only then salary is increased by 1000 INR and display name, employee id and increased salary of employee, if employee id is less than 201 then print “No Appraisal”

Sample Input 0

200
shruti
manager
Sample Output 0

No Appraisal