Create Day_28
Questions 1.
Write a program to print the names of students by creating a Student class. If instead of name some other data type is passed then the name should be "Unknown", otherwise the name should be equal to the String value passed while creating object of Student class.

Input Format

Rahul 33

Constraints

Use constructor with argument to initialize the name Create an object with name to print the name of the student Pass Two inputs such as one is name and the other with different data type

Output Format

Rahul Unknown

Sample Input 0

Rahul
33
Sample Output 0

Rahul
Unknown
Sample Input 1

Rahul
&
Sample Output 1

Rahul
Unknown

Questions 2.
Consider that in a plantation activity, the municipality is planting trees in an order such that the smallest tree is planted first and the biggest at the end. While arranging the trees, before planting, to position them at their positions, they are swapped in such a way that a tree can only be swapped with its adjacent trees. wAP to implemennt the same using method which contain the logic of sorting. Consider 10 trees height as input.

Input Format

71 70 69 68 65 71 68 61 63

Constraints

Consider that there are 10 are present in the activity and their lenghts are in integer numbers (between 60 inches to 72 inches).

Output Format

61 63 65 68 68 69 70 71 71

Sample Input 0

71 71 70 69 68 65 78 88 61 63
Sample Output 0

61 63 65 68 69 70 71 71 78 88